Refactor role management: remove IRoleClient, integrate role retrieval into AccessPackageClient, and update RoleService to use AccessPackageClient for role fetching
