+++ backend/src/Altinn.AccessManagement.UI/Altinn.AccessManagement.UI.Core/Services/Interfaces/IRoleService.cs
@@ -8,17 +8,11 @@ namespace Altinn.AccessManagement.UI.Core.Services.Interfaces
         public interface IRoleService
         {
                 /// <summary>
-                ///     Performs a search for access packages based on the provided parameters and sorts them into a list of areas for frontend to display
-                /// </summary>
-                /// <param name="languageCode">languageCode.</param>
-                /// <returns>the resources that match the filters and search string corresponding to the provided page.</returns>
-                Task<List<Role>> GetAllRoles(string languageCode);
-
-                /// <summary>
                 ///     Gets the roles for a user
                 /// </summary>
                 /// <param name="languageCode">languageCode.</param>
-                /// <param name="userId">user id.</param>
+                /// <param name="rightOwnerUuid">the right owner to retrieve roles for</param>
+                /// <param name="rightHolderUuid">the right holder to retrieve roles for</param>
                 /// <returns></returns>
                 Task<List<RoleAssignment>> GetRolesForUser(string languageCode, Guid rightOwnerUuid, Guid rightHolderUuid);
     }
