+++ backend/src/Altinn.AccessManagement.UI/Altinn.AccessManagement.UI.Core/Services/RoleService.cs
@@ -11,7 +11,6 @@ namespace Altinn.AccessManagement.UI.Core.Services
     {
         private readonly IAccessManagementClient _accessManagementClient;
         private readonly IAccessPackageClient _accessPackageClient;
-        private readonly IRoleClient _roleClient;
         private readonly ILookupService _lookupService;
         private readonly JsonSerializerOptions options = new JsonSerializerOptions
         {
@@ -23,26 +22,18 @@ namespace Altinn.AccessManagement.UI.Core.Services
         /// </summary>
         /// <param name="accessManagementClient">The access management client.</param>
         /// <param name="accessPackageClient">The access package client.</param>
-        /// <param name="lookupService">The lookup service.</param>
-        /// <param name="roleClient">The role client.</param>
-        public RoleService(IAccessManagementClient accessManagementClient, IAccessPackageClient accessPackageClient, ILookupService lookupService, IRoleClient roleClient)
+        /// <param name="lookupService">The lookup service.</param>s
+        public RoleService(IAccessManagementClient accessManagementClient, IAccessPackageClient accessPackageClient, ILookupService lookupService)
         {
             _accessManagementClient = accessManagementClient;
             _accessPackageClient = accessPackageClient;
             _lookupService = lookupService;
-            _roleClient = roleClient;
-        }
-
-        /// <inheritdoc />
-        public async Task<List<Role>> GetAllRoles(string languageCode)
-        {
-            return await _roleClient.GetRoles(languageCode);   
         }
 
         /// <inheritdoc />
         public async Task<List<RoleAssignment>> GetRolesForUser(string languageCode, Guid rightOwnerUuid, Guid rightHolderUuid)
         {
-            return await _roleClient.GetRolesForUser(languageCode, rightOwnerUuid, rightHolderUuid);
+            return await _accessPackageClient.GetRolesForUser(languageCode, rightOwnerUuid, rightHolderUuid);
         }
     }
 }
