--- backend/src/Altinn.AccessManagement.UI/Altinn.AccessManagement.UI.Integration/Clients/RoleClient.cs
@@ -1,104 +1,0 @@
-﻿using System.Net;
-using System.Text.Json;
-using Altinn.AccessManagement.UI.Core.ClientInterfaces;
-using Altinn.AccessManagement.UI.Core.Extensions;
-using Altinn.AccessManagement.UI.Core.Helpers;
-using Altinn.AccessManagement.UI.Core.Models.AccessPackage;
-using Altinn.AccessManagement.UI.Core.Models.Role;
-using Altinn.AccessManagement.UI.Core.Services.Interfaces;
-using Altinn.AccessManagement.UI.Integration.Configuration;
-using Microsoft.AspNetCore.Http;
-using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Options;
-
-namespace Altinn.AccessManagement.UI.Integration.Clients
-{
-    /// <summary>
-    /// Client for interacting with Access packages
-    /// </summary>
-    public class RoleClient : IRoleClient
-    {
-        private readonly ILogger _logger;
-        private readonly HttpClient _client;
-        private readonly IHttpContextAccessor _httpContextAccessor;
-        private readonly PlatformSettings _platformSettings;
-        private readonly IAccessTokenProvider _accessTokenProvider;
-        private readonly JsonSerializerOptions _serializerOptions = new JsonSerializerOptions { PropertyNameCaseInsensitive = true };
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="RoleClient"/> class
-        /// </summary>
-        /// <param name="httpClient">the handler for httpclient service</param>
-        /// <param name="logger">the handler for logger service</param>
-        /// <param name="httpContextAccessor">the handler for httpContextAccessor service</param>
-        /// <param name="platformSettings"> platform settings configuration</param>
-        /// <param name="accessTokenProvider">the handler for access token generator</param>
-        public RoleClient(
-            HttpClient httpClient,
-            ILogger<RoleClient> logger,
-            IHttpContextAccessor httpContextAccessor,
-            IOptions<PlatformSettings> platformSettings,
-            IAccessTokenProvider accessTokenProvider)
-        {
-            _logger = logger;
-            _platformSettings = platformSettings.Value;
-            httpClient.BaseAddress = new Uri(_platformSettings.ApiAccessManagementEndpoint); // TODO: Change to final ApiRoleEndpoint
-            httpClient.DefaultRequestHeaders.Add(_platformSettings.SubscriptionKeyHeaderName, _platformSettings.SubscriptionKey);
-            _client = httpClient;
-            _httpContextAccessor = httpContextAccessor;
-            _accessTokenProvider = accessTokenProvider;
-        }
-
-        /// <inheritdoc />
-        public async Task<List<Role>> GetRoles(string languageCode)
-        {
-            try
-            {
-                var endpointUrl = "/role";
-                string token = JwtTokenUtil.GetTokenFromContext(_httpContextAccessor.HttpContext, _platformSettings.JwtCookieName);
-
-                HttpResponseMessage response = await _client.GetAsync(token, endpointUrl);
-                string responseContent = await response.Content.ReadAsStringAsync();
-                if (response.StatusCode == HttpStatusCode.OK)
-                {
-                    return JsonSerializer.Deserialize<List<Role>>(responseContent, _serializerOptions);
-                }
-
-                _logger.LogError("AccessManagement.UI // RoleClient // GetRoles // Unexpected HttpStatusCode: {StatusCode}\n {responseBody}", response.StatusCode, responseContent);
-                return null;
-            }
-            catch (Exception e)
-            {
-                _logger.LogError(e, "AccessManagement.UI // RoleClient // GetRoles // Exception");
-                return null;
-            }
-        }
-
-        /// <inheritdoc />
-        public async Task<List<RoleAssignment>> GetRolesForUser(string languageCode, Guid rightOwnerUuid, Guid rightHolderUuid)
-        {
-            try
-            {
-                var endpointUrl = $"/assignment?from={rightOwnerUuid}&to={rightHolderUuid}";
-                string token = JwtTokenUtil.GetTokenFromContext(_httpContextAccessor.HttpContext, _platformSettings.JwtCookieName);
-
-                HttpResponseMessage response = await _client.GetAsync(token, endpointUrl, languageCode);
-                string responseContent = await response.Content.ReadAsStringAsync();
-                if (response.StatusCode == HttpStatusCode.OK)
-                {
-                    return JsonSerializer.Deserialize<List<RoleAssignment>>(responseContent, _serializerOptions);
-                }
-                else
-                {
-                    _logger.LogError("AccessManagement.UI // RoleClient // GetRolesForUser // Unexpected HttpStatusCode: {StatusCode}\n {responseBody}", response.StatusCode, responseContent);
-                    return null;
-                }
-            }
-            catch (Exception e)
-            {
-                _logger.LogError(e, "AccessManagement.UI // RoleClient // GetRolesForUser // Exception");
-                return null;
-            }
-        }
-    }
-}
